fix(wine): restore gnome-session-xsession packages
Fedora Workstation no longer pre-installs the deprecated GNOME X11
session for new installations. The problem is that Wine can sometimes
end up in a broken state on Wayland (presumably, something wrong gets
cached somewhere), which causes games that previously ran fine to start
crashing every few minutes. It's not clear to me what the cause is or
where the faulty cache lies, but using X11 stops the crashes from
happening, so we’ll want to preserve the GNOME X11 session for now.