Update kitty theme configurations
Some CLI tools color reproduction capacity are framed or determined by
the underlying terminal emulator. Sometimes, they cannot produce a color
that lies outside the range of what the terminal's color palette renders
possible.

This is immediately noticeable when the terminal's first 16 codes do not
include a pure black value for the termcol0 entry and a pure white for
termcol15. They cannot set the correct background or foreground (inverse
of the background). It thus falls back to the closest approximation,
which is not pure.